refactor(configurator): split render profile page modules
